requirements.txt
pydantic==2.10.6
asyncpg==0.30.0
SQLAlchemy==2.0.38
alembic==1.15.1
pydantic-settings==2.8.1
bcrypt==4.3.0
passlib==1.7.4
Jinja2==3.1.6
sanic-ext==24.12.0
python-jose==3.4.0
email-validator==2.2.0